findResultArray integred withdrawService
