+++ routes/requestRouter.js
@@ -1,19 +1,19 @@
 /* eslint-disable no-throw-literal */
 const express = require('express')
 const router = express.Router()
-const mainFunction = require('../services/mainFunction')
+const withdrawService = require('../services/withdrawService')
 
 const availableNotes = [100, 50, 20, 10]
 
 router.get('/:amount', (req, res, next) => {
-  mainFunction(req.params.amount, availableNotes, res)
+  withdrawService(req.params.amount, availableNotes, res)
 })
 
 router.get('/', (req, res, next) => {
-  mainFunction(req.query.amount, availableNotes, res)
+  withdrawService(req.query.amount, availableNotes, res)
 })
 
 router.post('/', (req, res, next) => {
-  mainFunction(req.body.amount, availableNotes, res)
+  withdrawService(req.body.amount, availableNotes, res)
 })
 module.exports = router
